+++ src/components/AppMenu.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
-import MenuItem from "@material-ui/core/MenuItem";
-import Menu from "@material-ui/core/Menu";
+import MenuItem from '@material-ui/core/MenuItem';
+import Menu from '@material-ui/core/Menu';
 
 import { WithLocale } from '../util/lang';
 import { I18N } from '../config';
@@ -13,21 +13,21 @@ interface AppMenuProps extends WithLocale {
 
 export default class AppMenu extends React.Component<AppMenuProps> {
 
-    handleLanguage = () => {
-        this.props.itemSelect("language");
-    };
-
-    handleServer = () => {
-        this.props.itemSelect("server");
-    };
-
-    handleSettings = () => {
-        this.props.itemSelect("settings");
-    };
-
     handleClose = () => {
         this.props.itemSelect(null);
     };
+
+    handleLanguage = () => {
+        this.props.itemSelect('language');
+    };
+
+    handleServer = () => {
+        this.props.itemSelect('server');
+    };
+
+    // handleSettings = () => {
+    //     this.props.itemSelect('settings');
+    // };
 
     render() {
         const {anchorElement} = this.props;
@@ -40,7 +40,7 @@ export default class AppMenu extends React.Component<AppMenuProps> {
             >
                 <MenuItem onClick={this.handleLanguage}>{I18N.get('Language')}</MenuItem>
                 <MenuItem onClick={this.handleServer}>{I18N.get('Server')}</MenuItem>
-                <MenuItem onClick={this.handleSettings}>{I18N.get('Settings')}</MenuItem>
+                {/*<MenuItem onClick={this.handleSettings}>{I18N.get('Settings')}</MenuItem>*/}
             </Menu>
         );
     }